Fix 8 defects found by an adversarial review of M2.17/M2.18
A five-lens multi-agent review of the last three commits raised 22 findings; 10
survived independent refutation. Eight are fixed here, each proven in both
directions where a test could hold it.

THE WORST ONE WAS MINE, AND IT WOULD HAVE WEDGED TRACKS.
bumpClipVersionFor advanced the GLOBAL counter first and the per-track counter
second. The publisher GATES on the global and PUBLISHES the per-track value —
so a publish landing between the two increments latched the new gate while
writing the OLD per-track version, then returned early forever after because
the gate already matched. That track's published base would be permanently one
behind, and every client reading it would present a stale base and have every
edit rejected. The window is widest in bumpAllTrackClipVersions, where the
global bump preceded a tracksMutex acquisition the publisher takes on every
iteration — so a whole all-tracks rebuild could complete inside it, and the
symptom would be a project that is wedged immediately after Load. Value first,
gate second, in both helpers. applyAddNote already had it right, which is what
made the inversion easy to miss.

AddTrack and RemoveTrack replaced a track's entire clip content and moved
NEITHER counter. The version-gated regions were never rebuilt, so a removed
track's notes stayed published; and because AddTrack reuses the tombstoned
TrackRuntime, a base read before the delete was still accepted against the
brand-new empty track in that slot. Both now bump.

daw-agent's add_notes still read the GLOBAL clip version as its base. daw-cli
was migrated in M2.17 and this was not — so the moment anyone edited another
track, every note the agent wrote would be silently rejected. Which is the
"agent works on track 4 while you type on track 1" case itself.

`daw-cli get clip` returned the PREVIOUS invocation's answer. It stamped every
request with the constant id 0x5ADD, and the read-back region is a single
persistent slot holding the last answer — so from the second call on it matched
the stale snapshot and returned before the engine had served the new request.
The reviewer's repro, confirmed here: asking for 2 bars right after 8 printed
the 8-bar window, and a `get clip` straight after a `do note` reported the note
ABSENT — exactly the observation that makes an agent conclude its write was
lost. The id is now unique per invocation (pid mixed with the clock), which
also stops two concurrent requesters taking delivery of each other's answers,
newly possible since `do` stopped needing --force. Verified against the same
repro: 30720000/7680000/30720000/7680000/15360000, and the note is visible on
the first read.

I BROKE LEVEL-MATCHED BYPASS while fixing the instrument meter. Widening the
output loop to numOutputs left the ratio comparing an output RMS over MORE
channels than the input RMS was taken over, so for any plugin narrower than its
bus outRms/inRms stopped being a gain. The meter still measures every channel;
the level-match ratio now accumulates the matched channel subset separately.

A negative Euclidean octave_offset never survived a save. The device-level
serializer cast the int8_t through uint32_t, writing 4294967295 for -1, which
the reader's get<int32_t> cannot translate — so boost returned the default and
the knob reset to 0 on every reload. The node-level serializer 180 lines up was
always correct; the two had drifted. project_file_tests now round-trips a
negative offset (and velocity/base_octave/degree, all of which were untested);
with the unsigned cast restored it reports 3 failures.

The -UNDEBUG loop that re-enabled 161 compiled-out assertions ran BEFORE
phase3_tests was declared, and `if (TARGET ...)` silently skips a target that
does not exist yet — so the largest test target was still building with
-DNDEBUG while the CMake read as though it were not. Moved to the end of the
file with a note saying why it must stay there.

AND THE METER CHECK COULD NOT SEE HALF THE BUG IT EXISTS FOR. Its fixture had
no clips, so the instrument received no MIDI and its out_* read SILENT whether
or not the output metering worked — reverting the output-loop bound left it
green. The fixture instrument now generates a continuous tone
(DAW_FAKE_TONE_HZ), which is measurable in every block instead of the ~9% that
contain a note pulse, and the check asserts out_peak is about -1200 mB for a
0.25-amplitude tone. With the loop bound reverted it now reads silent and
fails, as it always should have.

Full ctest 39/39. level_match_bypass_check, per_track_version_check,
multi_producer_ring_check, lane_quantize_check all pass.

+++ apps/project_file.cpp
@@ -624,8 +624,13 @@ std::string serializeProject(const ProjectDocument& document) {
         writer.key("duration_ticks",
                    static_cast<uint64_t>(device.euclideanConfig.duration_ticks));
         writer.key("degree", static_cast<uint32_t>(device.euclideanConfig.degree));
+        // int64_t, NOT uint32_t: octave_offset is int8_t, and casting a negative
+        // through unsigned wrote 4294967295 for -1. The reader's get<int32_t> cannot
+        // translate that, so boost returned the default and a negative octave offset
+        // silently became 0 on every reload. The node-level serializer 180 lines up
+        // always did this correctly; the two paths had drifted apart.
         writer.key("octave_offset",
-                   static_cast<uint32_t>(device.euclideanConfig.octave_offset));
+                   static_cast<int64_t>(device.euclideanConfig.octave_offset));
         writer.key("velocity", static_cast<uint32_t>(device.euclideanConfig.velocity));
         writer.key("base_octave",
                    static_cast<uint32_t>(device.euclideanConfig.base_octave));
